// SiteEsprit-main/esprit-backend/src/guards/roles.enum.ts
export enum adminRoles {
  admin = 'admin',
  addRoles = 'addRoles'
}

export enum clubRoles {
  createClub = 'createClub',
  updateClub = 'updateClub',
  deleteClub = 'deleteClub'
}

export enum partnershipRoles {
  createPartnership = 'createPartnership',
  updatePartnership = 'updatePartnership',
  deletePartnership = 'deletePartnership'
}

export enum pressRoles {
  createPress = 'createPress',
  updatePress = 'updatePress',
  deletePress = 'deletePress',
  handleArchivedPress = 'handleArchivedPress'
}

export enum EspritIngenieurRoles {
  createEspritIngenieur = 'createEspritIngenieur',
  updateEspritIngenieur = 'updateEspritIngenieur',
  deleteEspritIngenieur = 'deleteEspritIngenieur',
}


export enum ResultatCamerounRoles {
  createResultatCameroun = 'createResultatCameroun',
  updateResultatCameroun = 'updateResultatCameroun',
  deleteResultatCameroun = 'deleteResultatCameroun',
}


export enum rdiRoles {
  createRdi = 'createRdi',
  updateRdi = 'updateRdi',
  deleteRdi = 'deleteRdi'
}

export enum unitRoles {
  createUnit = 'createUnit',
  updateUnit = 'updateUnit',
  deleteUnit = 'deleteUnit',
  handleArchivedUnit = 'handleArchivedUnit'
}

export enum sliderRoles {
  createSlider = 'createSlider',
  updateSlider = 'updateSlider',
  deleteSlider = 'deleteSlider',
  handleArchivedSlider = 'handleArchivedSlider'
}

export enum eventRoles {
  createEvent = 'createEvent',
  updateEvent = 'updateEvent',
  deleteEvent = 'deleteEvent',
  handleArchivedEvent = 'handleArchivedEvent'
}

export enum timeTableRoles {
  createTimeTable = 'createTimeTable',
  updateTimeTable = 'updateTimeTable',
  deleteTimeTable = 'deleteTimeTable',
  handleArchivedTimeTable = 'handleArchivedTimeTable'
}

export enum faqRoles {
  createFaq = 'createFaq',
  updateFaq = 'updateFaq',
  deleteFaq = 'deleteFaq'
}

export enum teftefaqRoles {
  createTefTefaq = 'createTefTefaq',
  updateTefTefaq = 'updateTefTefaq',
  deleteTefTefaq = 'deleteTefTefaq',
  handleArchivedTefTefaq = 'handleArchivedTefTefaq'
}

export enum memberRoles {
  createMember = 'createMember',
  updateMember = 'updateMember',
  deleteMember = 'deleteMember',
  handleArchivedMember = 'handleArchivedMember'
}

export enum productionRoles {
  createProduction = 'createProduction',
  updateProduction = 'updateProduction',
  deleteProduction = 'deleteProduction',
  handleArchivedProduction = 'handleArchivedProduction'
}